requirements.txt
cryptography==2.7
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-formtools==2.1
django-itemlist>=0.2.2
django-model-utils==3.2.0
django-proxy==1.2.1
django-python3-ldap==0.11.2
jsonfield2==3.0.2
msgpack==0.6.1
numpy==1.17.1
pip-chill==0.1.8
